Rename project name to rucio-dataset-monitoring
