+++ src/go/rucio-dataset-monitoring/static/templates/datasets.tmpl
@@ -588,7 +588,7 @@
                     action: function (e, dt, node, config)
                     {
                         //This will send the page to the location specified
-                        window.open("https://github.com/dmwm/CMSMonitoring/tree/master/src/go/rucio-dataset-mon-go", "_blank");
+                        window.open("https://github.com/dmwm/CMSMonitoring/tree/master/src/go/rucio-dataset-monitoring", "_blank");
                     }
                 }
             ]
