+++ src/go/rucio-dataset-monitoring/static/templates/rse_detail_table.tmpl
